+++ ant/ant.c
@@ -8,6 +8,7 @@
 
 #include <stdlib.h>
 #include "ant.h"
+#include "../parser/parser.h"
 
 Ant* init_ant(int x_size, int y_size) {
     // allocate memory for the ant and set the
@@ -31,7 +32,7 @@ void rotate_ant(Ant* ant, int num_times, int dir) {
     }
 }
 
-void move_ant_forward(Ant* ant) {
+void move_ant_forward(Ant* ant, int x_size, int y_size) {
     // The cardinal directions N, E, S, W are represented with values 0, 1, 2, 3, respectively.
     // That means a value of 0 or 2 affects the y direction, and 1 or 3 affects the x direction.
     // Check which axis we need to change first.
@@ -41,18 +42,58 @@ void move_ant_forward(Ant* ant) {
         // W changes the x position relatively by -1.
         // 2 - dir will correctly map 1 ~> 1 and 3 ~> -1.
         ant->current_position.x += 2 - ant->direction;
+        
+        // Check if we are in cylinder/torus mode and apply a wrap
+        // if the ant is out of bounds on the x axis.
+        if(BORDER_MODE > 0 && BORDER_MODE <= 2) {
+            ant->current_position.x += ant->current_position.x <  0      ?  x_size :
+                                       ant->current_position.x >= x_size ? -x_size : 0;
+        }
+        
+        // Check if we are in reflect mode and apply a redirection
+        // if the ant is out of bounds on the x-axis.
+        if(BORDER_MODE == 3) {
+            if(ant->current_position.x < 0) {
+                ant->current_position.x = 0;
+                ant->direction = 1;
+            } else if(ant->current_position.x >= x_size) {
+                ant->current_position.x = x_size-1;
+                ant->direction = 3;
+            }
+        }
     } else {
         // direction is either N or S (0 or 2).
         // N changes the y position relatively by -1.
         // S changes the y position relatively by +1.
-        // dir - 1 will correctly map 0 ~> -1 and 2 -> 1.
+        // dir - 1 will correctly map 0 ~> -1 and 2 ~> 1.
         ant->current_position.y += ant->direction - 1;
+        
+        // Check if we are in torus mode and apply a wrap
+        // if the ant is out of bounds on the y-axis.
+        if(BORDER_MODE == 2) {
+            ant->current_position.y += ant->current_position.y <  0      ?  y_size :
+                                       ant->current_position.y >= y_size ? -y_size : 0;
+        }
+        
+        // Check if we are in reflect mode and apply a redirection
+        // if the ant is out of bounds on the y-axis.
+        if(BORDER_MODE == 3) {
+            if(ant->current_position.y < 0) {
+                ant->current_position.y = 0;
+                ant->direction = 2;
+            } else if(ant->current_position.y >= y_size) {
+                ant->current_position.y = y_size-1;
+                ant->direction = 0;
+            }
+        }
     }
 }
 
-char validate_ant_pos(Ant* ant, int x_size, int y_size) {
-    return ant->current_position.x >= 0 &&
-           ant->current_position.x < x_size &&
-           ant->current_position.y >= 0 &&
-           ant->current_position.y < y_size;
+char validate_ant_pos(Ant* ant, int x_size, int y_size, int sequence_number) {
+    int x_pass = ant->current_position.x >= 0 && ant->current_position.x < x_size;
+    int y_pass = ant->current_position.x >= 0 && ant->current_position.y < y_size;
+    int torus_pass = sequence_number <= 30000;
+    return BORDER_MODE == 1 ? y_pass     :
+           BORDER_MODE == 2 ? torus_pass :
+           BORDER_MODE == 3 ? torus_pass : x_pass && y_pass;
 }