+++ langtons_ant.c
@@ -7,6 +7,7 @@
 #include <stdlib.h>
 #include "ant/ant.h"
 #include "encode/encode.h"
+#include "parser/parser.h"
 
 #define BOARD_LENGTH 128
 
@@ -14,13 +15,14 @@ char buffer[BOARD_LENGTH][BOARD_LENGTH];
 
 int main(int argc, char *argv[])
 {
+    parse_args(argc, argv);
     start_encode(BOARD_LENGTH, BOARD_LENGTH, 50);
     Ant* ant = init_ant(_x_size, _y_size);
-    while(validate_ant_pos(ant, _x_size, _y_size)) {
+    while(validate_ant_pos(ant, _x_size, _y_size, _sequence_number)) {
         char pos_val = buffer[ant->current_position.y][ant->current_position.x];
         rotate_ant(ant, 1, pos_val == 1 ? -1 : 1);
         buffer[ant->current_position.y][ant->current_position.x] = !pos_val;
-        move_ant_forward(ant);
+        move_ant_forward(ant, _x_size, _y_size);
         next_frame((char*) buffer);
     }
     finish_encode();